requirements.txt
gradio==2.3.4
opencv-python==4.5.1
torch==1.9.0
Pillow==8.1.0
